Code refactored. TopSearchView(SearchField), TopToggleView(Place Filter Toggle Filters), FilterCarouselView have been collected in a component so that it can be reused.
