+++ src/ListView.jsx
@@ -4,6 +4,10 @@ import {Toolbar, ToolbarButton, Page, Button, BackButton, Icon, Segment, SearchI
 
 import LocalizedStrings from 'react-localization';
 
+import TopToggleView from './TopToggleView';
+import TopSearchView from './TopSearchView';
+import FilterCarouselView from './FilterCarouselView';
+
 export default class ListView extends React.Component {
   constructor(props) {
     super(props);
@@ -21,8 +25,6 @@ export default class ListView extends React.Component {
 
     let selectedCode = localStorage.getItem("code");
     let cache = JSON.parse(localStorage.getItem("items" + selectedCode));
-    let availCategories = this.makeAvailCategoryListFromItems(cache.items);
-    let filterCarouselItems = this.drawCategoryCarousel(availCategories, []);
     
     this.listItemHeight = 110;
  
@@ -31,8 +33,6 @@ export default class ListView extends React.Component {
       favorites: favorites,
       strings: strings,
       filtered: [],
-      filterCarouselItems: filterCarouselItems,
-      availCategories: availCategories,
       filterCarouselIndex: 0,
       searchString: "",
       sigunguCode: 0,
@@ -40,189 +40,13 @@ export default class ListView extends React.Component {
     };
   }
 
-  makeAvailCategoryListFromItems(items) {
-    var availCategorySet = new Set();
-
-    /* read available category from API and store them into a set*/
-    for(let i = 0; i < items.length; i++) {
-      let item = items[i];
-      let cat3 = item.cat3;
-      if(cat3 != null) {
-        let cat3Code = cat3._text;
-        availCategorySet.add(cat3Code)
-      }
-    }
-
-    let categories_kr = require('public/category/category_kr.json'); /* load all category pack (kr) */
-    let categories_en = require('public/category/category_en.json'); /* load all category pack (kr) */
-    let categoryAll = {key: "0", value: "전체"};
-    let allCategories = null;
-
-    let lang = localStorage.getItem("lang");
-    if(lang == 'kr') {
-      allCategories = categories_kr;
-    } else {
-      allCategories = categories_en;
-      categoryAll = {key: "0", value: "All"};
-    }
- 
-    let availCategories = [];
-    availCategories.push(categoryAll);
-
-    /* compare all category and set item and if the category is available one, 
-     * push it into availCategories */
-    for(let i = 0; i < allCategories.length; i++) {
-      let category = allCategories[i];
-      if(availCategorySet.has(category.key)) {
-        availCategories.push(category);
-      }
-    }
-
-    return availCategories;
-  }
-
-  drawCategoryCarousel(availCategories, filtered) {
-    /* make category filter buttons */
-    const filterButtonStyle = {
-      width: '22%', // draw 4 buttons on single carousel
-      margin: '1%',
-      height: '30px',
-      fontSize: '0.7em',
-      padding: '1px'
-    };
-
-    let buttonStyle = {
-      width: '2%',
-      padding: '0px',
-      height: '30px',
-      fontSize: '0.7em'
-    };
-
-    const innerDivStyle = {
-      textAlign: "center",
-      margin: '0%',
-      height: '30px'
-    };
-
-    /* prev button for carousel */
-    let backButton = (
-      <Button modifier='quiet' style={buttonStyle} onClick={this.goPrevFilterCarousel.bind(this)}>
-        <div style={innerDivStyle}>&lt;</div>
-      </Button>);
-
-    /* next button for carousel */
-    let nextButton =(
-      <Button modifier='quiet' style={buttonStyle} onClick={this.goNextFilterCarousel.bind(this)}>
-        <div style={innerDivStyle}>&gt;</div>
-      </Button>); 
-    
-    let filterCarouselItems = [];
-    // four buttons in a row, so make length / 4 size of carousel item
-    for(let i = 0; i < availCategories.length / 4; i++) {
-      // four buttons in a row
-      let start = i * 4;
-      let end = start + 4;
-      // four buttons will be grouped in a buttonGroup array and then stored into a carousel item.
-      let buttonGroup = [];
-      // add prev button at first
-      buttonGroup.push(backButton);
-      for(let j = start; j < end; j++) {
-        // if a item with index is not null or not undefined, make that into button with key.
-        if(availCategories[j] != null) {
-          let category = availCategories[j];
-          let modifier = 'outline'; // default inactive button style is outline.
-          if(filtered.length <= 0) {
-            if(j == 0) modifier = null;
-          }
-          else {
-            for(let fi = 0; fi < filtered.length; fi++) {
-              if(filtered[fi] == category.key) {
-                modifier = null;
-                break;
-              }
-            }
-          }
-
-          // the first button is button for no filter. so let us make this button in active.
-          let button = (
-            <Button key={category.key} style={filterButtonStyle} modifier={modifier}
-              onClick={this.toggleFilterStatus.bind(this, category.key)}>
-              <div style={innerDivStyle}>
-                {category.value}
-              </div>
-            </Button>);
-          buttonGroup.push(button);
-        } else {
-          // make dummy button. if there is no proper item.
-          let button = (<Button 
-            style={{width:'21%', margin: '1%', height: '30px'}} modifier='quiet' disabled={true}>
-            <div style={innerDivStyle}>
-            </div></Button>);
-          buttonGroup.push(button);
-        }
-      }
-      // add next button at first
-      buttonGroup.push(nextButton);
-      // make carousel item with grouped buttons.
-      let carouselItem = (<CarouselItem key={i}>
-        <div style={{textAlign: "left"}}>{buttonGroup}</div></CarouselItem>);
-      // and add it.
-      filterCarouselItems.push(carouselItem);
-    }
-
-    return filterCarouselItems;
-  }
-
   showMenu() {
     this.props.showMenu();
   }
 
-  goPrevFilterCarousel() {
-    let change = this.state.filterCarouselIndex - 1 < 0 ? 
-      Math.round(this.state.availCategories.length / 4) - 1 : 
-      this.state.filterCarouselIndex - 1;
-
-    this.setState({filterCarouselIndex: change});
-  }
-
-  goNextFilterCarousel() {
-    let change = this.state.filterCarouselIndex + 1 > Math.round(this.state.availCategories.length / 4) - 1 ? 
-      0 : 
-      this.state.filterCarouselIndex + 1;
-
-    this.setState({filterCarouselIndex: change});
-  }
-
-  toggleFilterStatus(key) {
-    let newFiltered = this.state.filtered.slice(); // copy the array
-    if(key == '0') {
-      // When 'All' button(key: 0) is clicked?
-      // Turn on 'All' button. And clear filter.
-      newFiltered = [];
-    }
-    else {
-      let indexToRemove = -1;
-      
-      for(let i = 0; i < newFiltered.length; i++) {
-        let filter = newFiltered[i];
-        if(filter == key) {
-          indexToRemove = i;
-          break;
-        }
-      }
-      if(indexToRemove == -1)
-      {
-        newFiltered.push(key);
-      } else {
-        newFiltered.splice(indexToRemove, 1); // remove item
-      }
-    }
-    
-    let filterCarouselItems = this.drawCategoryCarousel(this.state.availCategories, newFiltered);
-
+  toggleFilterStatus(newFilteredList) {
     this.setState({
-      filterCarouselItems: filterCarouselItems,
-      filtered: newFiltered});
+      filtered: newFilteredList});
   }
 
 
@@ -366,21 +190,8 @@ export default class ListView extends React.Component {
   }
 
   render() {
-    const centerDiv = {
-      textAlign: 'center'
-    };
-
-    const innerDiv = {
-      margin: '1%'
-    };
-    
     const hrStyle = {
       margin: '1px'
-    };
-
-    const searchIconSize = {
-      default: 24,
-      material: 22
     };
 
     const styleToolbar = {
@@ -389,39 +200,23 @@ export default class ListView extends React.Component {
       margin: '0px'};
 
     let fullWidth = window.innerWidth + "px";
- 
-    let filterCarousel = this.state.filterCarouselItems.length <= 0 ? "Loading..." : 
-      (<Carousel
-         style={{width: fullWidth}}
-         onPostChange={this.handleCategoryChange.bind(this)} 
-         index={this.state.filterCarouselIndex} 
-         autoScrollRatio={0.3} 
-         autoScroll overscrollable swipeable>
-         {this.state.filterCarouselItems}
-       </Carousel>);
-   
+
     return (
       <Page renderToolbar={this.renderToolbar.bind(this)}>
         <div style={styleToolbar}>
-          <div style={innerDiv}>
-            <Segment index={this.state.segmentIndex} 
-              onPostChange={this.handleAddressFilter.bind(this)} style={{ width: '55%' }}>
-              <Button>{this.state.strings.all}</Button>
-              <Button>{this.state.strings.seoguipo}</Button>
-              <Button>{this.state.strings.jeju}</Button>
-            </Segment>
-          </div>
-          <div style={innerDiv}>
-            <SearchInput
-              placeholder='Search...' onChange={this.handleSearchBox.bind(this)} />
-            <Button onClick={this.handleSearchButton.bind(this)} 
-              modifier='quiet' style={{margin: '2px', padding: '2px'}}>
-              <Icon icon='md-search' size={searchIconSize} />
-            </Button>
-          </div>
+          <TopToggleView index = {this.state.segmentIndex}
+            onPostChange = {this.handleAddressFilter.bind(this)}
+            strings = {this.state.strings} />
+          <TopSearchView onChange={this.handleSearchBox.bind(this)} 
+            onClick={this.handleSearchButton.bind(this)}/> 
           <div>
             <hr style={hrStyle}/>
-            {filterCarousel}
+            <FilterCarouselView 
+              width = {fullWidth}
+              strings = {this.state.strings} 
+              items = {this.state.items}
+              onFilterClicked = {this.toggleFilterStatus.bind(this)}
+            />  
             <hr style={hrStyle}/>
           </div>
         </div>
