+++ src/MapView.jsx
@@ -8,6 +8,9 @@ import MapContainer from './MapContainer';
 import DetailView from './DetailView';
 import ListView from './ListView';
 import Marker from './Marker';
+import TopToggleView from './TopToggleView';
+import TopSearchView from './TopSearchView';
+import FilterCarouselView from './FilterCarouselView';
 
 export default class MapView extends React.Component {
   constructor(props) {
@@ -44,9 +47,6 @@ export default class MapView extends React.Component {
       items: [],
       numOfDrawnItems: 0,
       placeCarouselItems: [],
-      availCategories: [],
-      filterCarouselIndex: 0,
-      filterCarouselItems: [],
       strings: strings,
       favorites: favorites,
       filtered: [],
@@ -80,11 +80,9 @@ export default class MapView extends React.Component {
         setTimeout(resolve, sleepTime, ret); // set some timeout to render page first
       }).then(function(result) {
         // success callback
-        let {availCategories, filterCarouselItems, placeCarouselItems, markers} = result; 
+        let {placeCarouselItems, markers} = result; 
         this_.setState({
           items: cache.items, 
-          availCategories: availCategories, 
-          filterCarouselItems: filterCarouselItems,
           placeCarouselItems: placeCarouselItems,
           markers: markers,
           numOfDrawnItem: placeCarouselItems.length,
@@ -275,151 +273,13 @@ export default class MapView extends React.Component {
     return {placeCarouselItems: placeCarouselItems, markers: markers}; 
   }
 
-  drawCategoryCarousel(availCategories, filtered) {
-    /* make category filter buttons */
-    const filterButtonStyle = {
-      width: '22%', // draw 4 buttons on single carousel
-      margin: '1%',
-      height: '30px',
-      fontSize: '0.7em',
-      padding: '1px'
-    };
-
-    let buttonStyle = {
-      width: '2%',
-      padding: '0px',
-      height: '30px',
-      fontSize: '0.7em'
-    };
-
-    const innerDivStyle = {
-      textAlign: "center",
-      margin: '0%',
-      height: '30px'
-    };
-
-    /* prev button for carousel */
-    let backButton = (
-      <Button modifier='quiet' style={buttonStyle} onClick={this.goPrevFilterCarousel.bind(this)}>
-        <div style={innerDivStyle}>&lt;</div>
-      </Button>);
-
-    /* next button for carousel */
-    let nextButton =(
-      <Button modifier='quiet' style={buttonStyle} onClick={this.goNextFilterCarousel.bind(this)}>
-        <div style={innerDivStyle}>&gt;</div>
-      </Button>); 
-    
-    let filterCarouselItems = [];
-    // four buttons in a row, so make length / 4 size of carousel item
-    for(let i = 0; i < availCategories.length / 4; i++) {
-      // four buttons in a row
-      let start = i * 4;
-      let end = start + 4;
-      // four buttons will be grouped in a buttonGroup array and then stored into a carousel item.
-      let buttonGroup = [];
-      // add prev button at first
-      buttonGroup.push(backButton);
-      for(let j = start; j < end; j++) {
-        // if a item with index is not null or not undefined, make that into button with key.
-        if(availCategories[j] != null) {
-          let category = availCategories[j];
-          let modifier = 'outline'; // default inactive button style is outline.
-          if(filtered.length <= 0) {
-            if(j == 0) modifier = null;
-          }
-          else {
-            for(let fi = 0; fi < filtered.length; fi++) {
-              if(filtered[fi] == category.key) {
-                modifier = null;
-                break;
-              }
-            }
-          }
-
-          // the first button is button for no filter. so let us make this button in active.
-          let button = (
-            <Button key={category.key} style={filterButtonStyle} modifier={modifier}
-              onClick={this.toggleFilterStatus.bind(this, category.key)}>
-              <div style={innerDivStyle}>
-                {category.value}
-              </div>
-            </Button>);
-          buttonGroup.push(button);
-        } else {
-          // make dummy button. if there is no proper item.
-          let button = (<Button 
-            style={{width:'21%', margin: '1%', height: '30px'}} modifier='quiet' disabled={true}>
-            <div style={innerDivStyle}>
-            </div></Button>);
-          buttonGroup.push(button);
-        }
-      }
-      // add next button at first
-      buttonGroup.push(nextButton);
-      // make carousel item with grouped buttons.
-      let carouselItem = (<CarouselItem key={i}>
-        <div style={{textAlign: "left"}}>{buttonGroup}</div></CarouselItem>);
-      // and add it.
-      filterCarouselItems.push(carouselItem);
-    }
-
-    return filterCarouselItems;
-  }
-
-
-  makeAvailCategoryListFromItems(items) {
-    var availCategorySet = new Set();
-
-    /* read available category from API and store them into a set*/
-    for(let i = 0; i < items.length; i++) {
-      let item = items[i];
-      let cat3 = item.cat3;
-      if(cat3 != null) {
-        let cat3Code = cat3._text;
-        availCategorySet.add(cat3Code)
-      }
-    }
-
-    let categories_kr = require('public/category/category_kr.json'); /* load all category pack (kr) */
-    let categories_en = require('public/category/category_en.json'); /* load all category pack (kr) */
-    let categoryAll = {key: "0", value: "전체"};
-    let allCategories = null;
-
-    let lang = this.state.strings.getLanguage();
-    if(lang == 'kr') {
-      allCategories = categories_kr;
-    } else {
-      allCategories = categories_en;
-      categoryAll = {key: "0", value: "All"};
-    }
- 
-    let availCategories = [];
-    availCategories.push(categoryAll);
-
-    /* compare all category and set item and if the category is available one, 
-     * push it into availCategories */
-    for(let i = 0; i < allCategories.length; i++) {
-      let category = allCategories[i];
-      if(availCategorySet.has(category.key)) {
-        availCategories.push(category);
-      }
-    }
-
-    return availCategories;
-  }
-
   makeContents(items, favorites, sigunguCode) {
     let emptyFilter = [];
 
-    let availCategories = this.makeAvailCategoryListFromItems(items);
-    let filterCarouselItems = this.drawCategoryCarousel(availCategories, emptyFilter);
     let {placeCarouselItems, markers} = 
       this.makeItemCarouselAndMarkers(items, favorites, emptyFilter, sigunguCode, "");
 
     return {
-      availCategories: availCategories,
-      filterCarouselItems: filterCarouselItems,
       placeCarouselItems: placeCarouselItems,
       markers: markers
     };
@@ -449,13 +309,11 @@ export default class MapView extends React.Component {
         let items = this_.readItemsFromResponseText(xhr.responseText);
         this_.writeItemCache(code, items); // write cache to manage # of calls of API
 
-        let {availCategories, filterCarouselItems, placeCarouselItems, markers} = 
+        let {placeCarouselItems, markers} = 
           this_.makeContents(items, this_.state.favorites, this_.state.sigunguCode);
 
         this_.setState({
           items: items, 
-          availCategories: availCategories, 
-          filterCarouselItems: filterCarouselItems,
           placeCarouselItems: placeCarouselItems,
           markers: markers,
           numOfDrawnItem: placeCarouselItems.length,
@@ -471,58 +329,17 @@ export default class MapView extends React.Component {
     });
   }
 
-  toggleFilterStatus(key) {
-    let newFiltered = this.state.filtered.slice(); // copy the array
-    if(key == '0') {
-      // When 'All' button(key: 0) is clicked?
-      // Turn on 'All' button. And clear filter.
-      newFiltered = [];
-    }
-    else {
-      let indexToRemove = -1;
-      
-      for(let i = 0; i < newFiltered.length; i++) {
-        let filter = newFiltered[i];
-        if(filter == key) {
-          indexToRemove = i;
-          break;
-        }
-      }
-      if(indexToRemove == -1)
-      {
-        newFiltered.push(key);
-      } else {
-        newFiltered.splice(indexToRemove, 1); // remove item
-      }
-    }
-    
-    let filterCarouselItems = this.drawCategoryCarousel(this.state.availCategories, newFiltered);
-    let {placeCarouselItems, markers} = 
+  toggleFilterStatus(newFilteredList) {
+   let {placeCarouselItems, markers} = 
       this.makeItemCarouselAndMarkers(
-        this.state.items, this.state.favorites, newFiltered, this.state.sigunguCode, this.state.searchString);
+        this.state.items, this.state.favorites, newFilteredList, 
+        this.state.sigunguCode, this.state.searchString);
 
     this.setState({
-      filterCarouselItems: filterCarouselItems,
-      filtered: newFiltered,
+      filtered: newFilteredList,
       placeCarouselItems: placeCarouselItems,
       markers: markers,
       numOfDrawnItem: placeCarouselItems.length});
-  }
-
-  goPrevFilterCarousel() {
-    let change = this.state.filterCarouselIndex - 1 < 0 ? 
-      Math.round(this.state.availCategories.length / 4) - 1 : 
-      this.state.filterCarouselIndex - 1;
-
-    this.setState({filterCarouselIndex: change});
-  }
-
-  goNextFilterCarousel() {
-    let change = this.state.filterCarouselIndex + 1 > Math.round(this.state.availCategories.length / 4) - 1 ? 
-      0 : 
-      this.state.filterCarouselIndex + 1;
-
-    this.setState({filterCarouselIndex: change});
   }
 
   prevItem() {
@@ -573,10 +390,6 @@ export default class MapView extends React.Component {
    this.props.navigator.pushPage({ 
       component: ListView
     });
-  }
-
-  handleCategoryChange(e) {
-    this.setState({filterCarouselIndex: e.activeIndex});
   }
 
   handlePlaceChange(e) {
@@ -644,26 +457,11 @@ export default class MapView extends React.Component {
       textAlign: 'center'
     };
 
-    const innerDiv = {
-      margin: '1%'
-    };
-    
     const hrStyle = {
       margin: '1px'
     };
 
-
     let fullWidth = window.innerWidth + "px";
-
-    let filterCarousel = this.state.filterCarouselItems.length <= 0 ? "Loading..." : 
-      (<Carousel
-         style={{width: fullWidth}}
-         onPostChange={this.handleCategoryChange.bind(this)} 
-         index={this.state.filterCarouselIndex} 
-         autoScrollRatio={0.3} 
-         autoScroll overscrollable swipeable>
-         {this.state.filterCarouselItems}
-       </Carousel>);
 
     let placeCarousel = this.state.items.length <= 0 ? (<ProgressCircular indeterminate />) :
       (<Carousel
@@ -716,36 +514,25 @@ export default class MapView extends React.Component {
         {markers}
       </MapContainer>);
 
-    const searchIconSize = {
-      default: 24,
-      material: 22
-    };
-
     return (
-      <Page renderToolbar={this.renderToolbar.bind(this)}>
-        <div style={centerDiv}>
-          <div style={innerDiv}>
-            <Segment index={this.state.segmentIndex} 
-              onPostChange={this.handleAddressFilter.bind(this)} style={{ width: '55%' }}>
-              <Button>{this.state.strings.all}</Button>
-              <Button>{this.state.strings.seoguipo}</Button>
-              <Button>{this.state.strings.jeju}</Button>
-            </Segment>
-          </div>
-          <div style={innerDiv}>
-            <SearchInput
-              placeholder='Search...' onChange={this.handleSearchBox.bind(this)} />
-            <Button onClick={this.handleSearchButton.bind(this)} 
-              modifier='quiet' style={{margin: '2px', padding: '2px'}}>
-              <Icon icon='md-search' size={searchIconSize} />
-            </Button>
-          </div>
-          <div style={{marginTop: '1%', marginBottom: '1%'}}>
+      <Page renderToolbar = {this.renderToolbar.bind(this)}>
+        <div style = {centerDiv}>
+          <TopToggleView index = {this.state.segmentIndex} 
+            onPostChange = {this.handleAddressFilter.bind(this)} 
+            strings = {this.state.strings}/>
+          <TopSearchView onChange = {this.handleSearchBox.bind(this)} 
+            onClick = {this.handleSearchButton.bind(this)}/> 
+          <div style = {{marginTop: '1%', marginBottom: '1%'}}>
             {map}
           </div>
           <div>
-            <hr style={hrStyle}/>
-            {filterCarousel}
+            <hr style = {hrStyle}/>
+            <FilterCarouselView 
+              width = {fullWidth}
+              strings = {this.state.strings} 
+              items = {this.state.items}
+              onFilterClicked = {this.toggleFilterStatus.bind(this)}
+            />
             <hr style={hrStyle}/>
           </div>
           <div>
